fix: move community utility hooks to separate file
- Create src/hooks/useCommunityHooks.ts for utility hooks
- Move useCommunityContext, useSelectedCommunity, useCommunityPermissions
- Update all imports across 10 files to use new hook module
- Reduces React Fast Refresh warnings in CommunityContext.tsx
