+++ src/components/dashboard/CommunitySelector.tsx
@@ -9,7 +9,7 @@ import {
 import { Badge } from "@/components/ui/badge";
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { Button } from "@/components/ui/button";
-import { useCommunityContext } from "@/contexts/CommunityContext";
+import { useCommunityContext } from "@/hooks/useCommunityHooks";
 import { Users, User, Crown, Shield, Plus } from "lucide-react";
 import { useNavigate } from "react-router-dom";
